gate-controller/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pyserial==3.5
requests==2.31.0
